refactor(mobile): decompose fizruk dualWrite adapter by operation family
Split the 804-LOC monolithic adapter.ts into per-family dispatch files
(workouts, customExercises, measurements, dailyLog, templates, programs,
wellbeing, activeWorkout) plus a shared _helpers.ts for numeric coercions.
adapter.ts is now a thin orchestrator (~160 LOC) that imports from each
family file and preserves the identical public API surface.
